* kill.cc (forcekill): Use unsigned int printf format specifier and 	cast to avoid compiler warnings where DWORD value is used. 	* regtool.cc: Ditto, throughout.

+++ winsup/utils/kill.cc
@@ -176,8 +176,8 @@ forcekill (int pid, int sig, int wait)
   if (!wait || WaitForSingleObject (h, 200) != WAIT_OBJECT_0)
     if (sig && !TerminateProcess (h, sig << 8)
 	&& WaitForSingleObject (h, 200) != WAIT_OBJECT_0)
-      fprintf (stderr, "%s: couldn't kill pid %u, %lu\n",
-	       prog_name, (unsigned) dwpid, GetLastError ());
+      fprintf (stderr, "%s: couldn't kill pid %u, %u\n",
+	       prog_name, (unsigned) dwpid, (unsigned int) GetLastError ());
   CloseHandle (h);
 }
